requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
Pillow>=8.0.0

